% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_mc.R
\name{baseline_mc}
\alias{baseline_mc}
\title{Montecarlo baseline data}
\usage{
baseline_mc(data = fitR2, montecarlo = 10000)
}
\arguments{
\item{data}{processed data}

\item{montecarlo}{number of repeated samples for accept-reject algorithm}
}
\value{
data after doing montecarlo
}
\description{
Montecarlo baseline data
}
