% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ci.R
\name{cal_ci}
\alias{cal_ci}
\title{Confident interval calculation}
\usage{
cal_ci(data, ci = 0.95, boot = T)
}
\arguments{
\item{data}{- input data}

\item{ci}{- percentage of confident interval}

\item{boot}{- doing bootstrapping}
}
\value{
confident interval
}
\description{
Confident interval calculation
}
