% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen_and_scale.R
\name{cen_and_scale}
\alias{cen_and_scale}
\title{Center and scale time variable}
\usage{
cen_and_scale(time)
}
\arguments{
\item{time}{- time variables}
}
\value{
data frame with scaled time variabels with mean and standard deviation
}
\description{
Center and scale time variable
}
