% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_form.R
\name{g_form}
\alias{g_form}
\title{G-formula}
\usage{
g_form(data, ii = 2, length = 12, am = 1, ay = 0)
}
\arguments{
\item{data}{- processed data}

\item{ii}{- id of patient}

\item{length}{- length of follow-up}

\item{am}{- am}

\item{ay}{- ay}
}
\value{
estimated dataframe
}
\description{
G-formula
}
