% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tvmedg}
\alias{plot.tvmedg}
\title{plot function}
\usage{
\method{plot}{tvmedg}(x, what = c("all", "cumY", "tvY"), ...)
}
\arguments{
\item{x}{data}

\item{what}{3 type of plot: cumY, tvY, or all}

\item{...}{arbitrary params}
}
\value{
plot
}
\description{
plot function
}
