% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(
  fix,
  expo,
  med,
  tvar,
  lag,
  outc,
  time,
  norev = NULL,
  LM = FALSE,
  data
)
}
\arguments{
\item{fix}{- time-fixed variables}

\item{expo}{- exposure variables}

\item{med}{- mediator variables}

\item{tvar}{- time-varying variables}

\item{lag}{- lag effect}

\item{outc}{- out come variables}

\item{time}{- time variables}

\item{norev}{- non-reversible variables}

\item{LM}{- L before M assumption}

\item{data}{- input data}
}
\value{
processed data with lag for each variables
}
\description{
Process data
}
