% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{resamp}
\alias{resamp}
\title{resampling data}
\usage{
resamp(data, boot = FALSE)
}
\arguments{
\item{data}{- processed data}

\item{boot}{- doing bootstrap}
}
\value{
data after resampling
}
\description{
resampling data
}
