% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFunc.R
\name{rFunc}
\alias{rFunc}
\title{predict binary variables}
\usage{
rFunc(mod, ndat)
}
\arguments{
\item{mod}{- model}

\item{ndat}{- new data}
}
\value{
output variables
}
\description{
predict binary variables
}
