% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvmedg.R
\name{tvmedg}
\alias{tvmedg}
\title{Time-varying mediation analysis using g-formular}
\usage{
tvmedg(
  data,
  fix,
  expo,
  med,
  tvar,
  lag,
  outc,
  time,
  norev = NULL,
  LM = FALSE,
  boot = FALSE,
  seed = 0,
  mreg = "binomial",
  lreg = c("binomial", "gaussian", "gaussian"),
  yreg = "binomial",
  dof = 3,
  montecarlo = 10000,
  length = 12,
  parallel = TRUE,
  nboot = 1,
  ci = 0.95
)
}
\arguments{
\item{data}{- input data}

\item{fix}{- time-fixed variables}

\item{expo}{- exposure variables}

\item{med}{- mediator variables}

\item{tvar}{- time-varying variables}

\item{lag}{- lag effect}

\item{outc}{- outcome variable}

\item{time}{- time variable}

\item{norev}{- non-reversible variables}

\item{LM}{- L before M assumption}

\item{boot}{- doing bootstraping}

\item{seed}{- set seed}

\item{mreg}{- mediator regression model}

\item{lreg}{- time-varying regression model}

\item{yreg}{- outcome regression model}

\item{dof}{- degree of freedom}

\item{montecarlo}{- number of repeated samples for accept-reject algorithm}

\item{length}{- length of follow-up}

\item{parallel}{- doing parrallel caculation}

\item{nboot}{- number of bootstrapping}

\item{ci}{- percentage of confidence interval}
}
\value{
q11,q10,q00
}
\description{
Time-varying mediation analysis using g-formular
}
